connection-editor: don't ellipsize unauthenticated CEPolkitButtons
Having CEPolkitButtons be sometimes ellipsized and sometimes not
causes the window width to change under some localizations.

After some discussion, it was decided that having to deal with an
authentication dialog does not really constitute the kind of
"additional input" that ellipses are meant to indicate, so just remove
them.

https://bugzilla.gnome.org/show_bug.cgi?id=642688