bluetooth: fix ModemManager1 code
The ModemManager1 code was assuming that the modem object would not be
created until after the Bluez Connect call returned, but that's only
true if this is the first time you tried to use it in the session. Fix
it.

NOTE: The MM 0.7 code is presumably also broken in the same way, but
is not fixed by the patch.