applet: don't show PIN unlock dialog before keyring search is done
Possible mis-merge of 8f49c2c7.  If the modem doesn't have a device
identifier or required unlock isn't a PIN, the second call to
unlock_dialog_new() would have no effect because unlock_dialog_new()
just returns if a dialog is already showing.  But if the PIN could
be in the keyring, the search would be started and the dialog would
be shown anyway, which is wrong.