utils: fix editor crash when context-menu's "Paste" item is used (rh #1089199)
When a user used "Paste" item from right-click context menu to copy a password
to "Password" GtkEntry in "Wi-Fi Security" tab, the editor crashed.
The reason was that we supposed that "pasword storage" menu is the only menu
attached to the entry. But there can be other menus attached, like the
right-click context menu. We fix that by marking "password storage" popup menu
to be able to identify it and not to mix up the menus.

https://bugzilla.redhat.com/show_bug.cgi?id=1089199