applet: don't check for gnome shell (bgo #707953)
The applet is no  longer autostarted under GNOME Shell, so it's pointless
to keep the shell watcher code around.  The only reason it was autostarted
was to handle VPN secrets (which GNOME Shell 3.0+ now does) and to auto-
unlock WWAN devices, which GNOME Shell does not yet do.  It has been
determined that should be handled differently, regardless of whether or
not GNOME Shell is running, since the auto-unlocking behavior really
annoys some users.

Also, the check only cares about GNOME Shell < 3.0, and 3.0 was released
in early 2011.  It's highly unlikely that anyone is using such an old
version of GNOME Shell with the latest nm-applet.

https://bugzilla.gnome.org/show_bug.cgi?id=707953

(commit message by dcbw)