applet: hardcode AP strength icons to 24x24 (bgo #741184)
The icon size in the menu is currently the same as the size of the
status icon.  That means if the status icon size is larger than
24x24 the menu looks awful.

We should be using GTK_ICON_SIZE_MENU here, but that's actually
a bit too small.  Since GTK hardcodes GTK_ICON_SIZE_MENU to 16x16
anyway, we might as well hardcode the size we want to use too.

https://bugzilla.gnome.org/show_bug.cgi?id=741184