libnm-gtk: fix library gettext usage (bgo #741215)
Libraries need to include <gi18n-lib.h>, not <gi18n.h>, so that _()
will get defined to "dgettext (GETTEXT_DOMAIN, string)" rather than
"gettext (string)" (which will use the program's default domain, which
works fine for programs in the network-manager-applet tree, but not for
external users). Likewise, we need to call bindtextdomain() so that
gettext can find the translations if the library is installed in a
different prefix from the program using it (and
bind_textdomain_codeset(), so it will know the translations are in
UTF-8 even if the locale isn't).

(based on NM commit 53f5e9afa466b0b1517cbac208ed01ab8f4b6274)

https://bugzilla.gnome.org/show_bug.cgi?id=741215