editor: remove "Network Type" combo from mobile broadband page
The network type was never used with ModemManager 1.0+, and NM itself
hasn't supported anything except ModemManager 1.0+ since 0.9.8.

Changing the network type is often a very heavy operation for the
modem, which may involve re-initialization or hardware reset, and
is not an appropriate operation to perform when connecting to
a network.  It should be an out-of-band action.