editor: improve color-indication for bad addresses and routes (bgo #660915)
Incorrect values are already indicated while they are edited. However, this
commit also makes the incorrect values visible afterwards by setting cell
background to red. It uses the pango markup which has the advantage of being
visible even if the row is selected. When there's no value in the cell, the
cell-background property is used instead.

https://bugzilla.gnome.org/show_bug.cgi?id=660915