editor: use ifname instead of UUID in slaves' master property (rh #1083186)
NetworkManager accepts both UUID and interface name in slave's master property
(identifying master). UUID is easier to use in the editor and may be more
appropriate for 'connection based' configurations. On the other hand, some
users may expect master property to be rather interface name. And this is also
the only supported option for some legacy configurations, such as initscripts
in Red Hat based distros.

We only use interface name for newly added slaves to ensure we don't change
existing connections (which would be unexpected for users).

https://bugzilla.redhat.com/show_bug.cgi?id=1083186