applet: don't crash if an agent is already registered
GNOME 3.16 makes it possible to run the legacy applets again, making it
possible to develop & debug the applet there. Unfortunately, we crash -- GNOME
registers its own secret agent.

Moreover the error handling in libnm (as of post-1.0 master) is broken. A
separate commit to daemon will address that. For now let's just avoid crashing
and produce a warning message.