applet: make new auto connections only available for current user (rh #1176042)
When a user clicks a new Wi-Fi network in nm-applet, AddAndActivateConnection()
D-Bus method is called to create a new connection and activate it. The
connection is available to all users by default (no connection.permissions).
However, this causes troubles when polkit action for modifying system
connections [1] is set to require admin access. Then a normal user cannot
connect to a new network, because it would require admin password to save
the new connection and finish AddAndActivateConnection(). We solve that by
creating new connections as private to current user only.

[1] org.freedesktop.NetworkManager.settings.modify.system set to auth
(i.e. auth_admin, auth_admin_keep)

https://bugzilla.redhat.com/show_bug.cgi?id=1176042