applet: drop bond, bridge, infiniband, team, and vlan support
At one point we thought people would want/need to control their
"enterprisey" devices via nm-applet, but nmcli (or static
autoconnect=true configuration) turns out to be a better answer
generally.

Additionally, probably the most common case of bridges currently is
when you are using virtualization or containers, and in that case the
bridge is handled entirely by the virt/container system, and showing
it in nm-applet just clutters things up, pushing Wi-Fi and VPN down to
the bottom of the menu.

So drop support for those devices.

https://bugzilla.gnome.org/show_bug.cgi?id=753369