libnma: fix wrong mnemonics in tooltips for password entries
The mnemonics of radio buttons *inside* the pop-up menu don't work
with respect to the whole window. They only work, if you already
opened the pop-up menu, and then press the mnemonic key.
But since you cannot open the pop-up menu with a shortcut,
this becomes moot and the user has to click anyway.

But more importantly, we are using the same string as tooltip
for the icon, which wrongly showed the underscore of the mnemonic.
Fix that by just removing the the mnemonics.

https://bugzilla.gnome.org/show_bug.cgi?id=754726

Fixes: 50b9650cc205a84dfe7082eb92e95ea2aaf554e8
(cherry picked from commit d888f571e5de9570c62b8ec557d8219e64ec45ba)