c-e: forward the validation error to print more sensible error message to stdout
When a connection doesn't verify, the "Save" dialog stays desensitized.
Even for an advanced user it's not immediately clear what causes the
valdation failure.

We already print a message like "Invalid setting Ethernet" to stdout.
Extend the message to also print an more detailed error that gets propagated
from the validation.

This later should be improved further to indicating the failure reason
in the UI too.