applet: use fewer deprecated GtkImageMenuItems (bgo #740574)
GtkImageMenuItem is deprecated and we should use fewer of them.  Most
GTK-based desktop environments seem to have stopped showing icons in
menus by default anyway.  This will allow us to simplify the code
for both the regular applet and the appindictor version.  So:

1) remove the stock icons from the static menu items so they won't
be displayed
2) override gtk-menu-images to TRUE for later appindicator changes,
since we actually want our custom menu item images displayed, just
not the stock ones (that we just removed anyway)