applet: fix crash due to wrongly handling references for icons
nma_icon_check_and_load() would not take an additional reference
for the @fallback_icon.

Note that nma_icons_free() would not unref the icon if it equals the
fallback_icon, but destroying @icon_cache would. So, during finalize()
we hit the bug, but also when nma_icons_reload() replaces
@fallback_icon.

https://bugzilla.gnome.org/show_bug.cgi?id=762421
https://bugzilla.redhat.com/show_bug.cgi?id=1284671
https://bugzilla.redhat.com/show_bug.cgi?id=1279937