applet: anticipate nma_icon_check_and_load() returning no icon
nma_icon_check_and_load() uses the fallback_icon when it fails
to load the icon.

Probably a fallback_icon should be always present. But nma_icons_reload()
goes through some lengths to cope with an error creating the
fallback_icon. So, either we assert that the fallback_icon is always
present, or users of nma_icon_check_and_load() must cope with a NULL
icon. Do the latter.