c-e: always add VPN connections to list even if there are no VPN plugins
Currently, when the user has no plugins installed all VPN connections will
be hidden.

But if at least one plugin is present, *all* VPN connections are shown --
regadless their type. For consistency, we should either always show all
VPN connections (even if we are unable to edit them) or we should only
show those for which we have a plugin present.

Do the former, as it is useful for the user to see that he has some
connections.