applet: remove assert(s_con) from applet_get_active_vpn_connection()
This assertion is just to strict. On client side, we cannot expect that
all connections that NetworkManager exposes are valid.

In libnm/libnm-glib there was an old bug, that connections which don't verify
could be completely bogus and thus hit this assertion. This got fixed by
https://cgit.freedesktop.org/NetworkManager/NetworkManager/commit/?id=23136ecf89f279479337ead355b7ff5e80465a0b
and with a fixed libnm/libnm-glib, we would always expect that a
connection has at least a NMSettingConnection instance.

Nonetheless, there is no reason to enforce that with an assert.

This was already fixed before by 22468c05291d3d88ddc68f8983bffe54f29f5f82
and reverted again by c86d66632ad76aa4a93593a07a9b4e25f18efbf7.

https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=815668
https://bugzilla.redhat.com/show_bug.cgi?id=1313866
https://bugzilla.redhat.com/show_bug.cgi?id=1314650
https://bugzilla.gnome.org/show_bug.cgi?id=765830