wifi-dialog: fix enabling of "Connect" button
When secrets for a connection become available we update all the
WirelessSecurity items, triggering multiple calls to
stuff_changed_cb() (since it is registered as a callback for changed
events). But only the currently selected security item should cause a
change in the activation of the "Connect" button.

https://bugzilla.gnome.org/show_bug.cgi?id=597839
(cherry picked from commit 093657df1478002f3d616717eb3a0ac4efae87aa)