c-e: support multiple "Add VPN connection" entires per service-type
For one given service-type, the UI shall be able to show multiple
entires in the list for creating a new connection.

This works by asking the plugin whether it supports multiple
"add-details". And if so, create an entry for each add-detail.

When adding a VPN connection, the editor creates a NMConnection
with initializing vpn.service-type and setting the add-detail to a
key according to the plugin's instruction.

For example, nm-openvpn supports multiple "vpn.connection-type" such as
"tls", "static-key", "password", "password-tls". With this change, it could
generate an entry for each type in nm-c-e's connection-add list. To do
that, it should announce add-details to be "tls", "static-key",
"password", "password-tls", and the "add-detail-key" to be "connection-type".

It also works to specify the add-detail on command line to create a
particular VPN connection:

  $ nm-connection-editor -c -t vpn:<SERVICE_TYPE>:<ADD_DETAIL>

For example:

  $ nm-connection-editor -c -t vpn:openconnect:nc