shared: remove header file "nm-dbus-compat.h"
This header was only used to have defines for well-known D-Bus paths.
These paths are an essential part of D-Bus, there is no need for a
define of those.

Often, we prefer defines over plain strings so that we can easier grep
for the users. E.g. instead of grepping for '"mac-address"', searching
for 'NM_SETTING_BRIDGE_MAC_ADDRESS' yields more precisely what we
search. That is not the case for these defines, as we probably won't
search for 'DBUS_SERVICE_DBUS'.