c-e: don't set the Cancel's button in "Add connection" dialog has_default
Previously, when selecting the "Create" button with the keyboard (tab key),
pressing ENTER would still invoke the "Cancel" action.

We either want that ENTER invokes the "Create" action, or that ENTER
invokes the currently selected button (be it "Create" or "Cancel").

This patch changes the UI to do the latter.

https://bugzilla.redhat.com/show_bug.cgi?id=1339565