nm-connection-editor: UI file for Proxy Tab
UI file for 'Proxy' tab contain entries/widgets for the three mode
nm-c-e is offering. On setting mode 'None' nothing is pushed inside
PacRunner (exception being IFname & Domains sent for all modes).
In mode 'Automatic' Pac Url or/and Pac Script can be provided
(override DHCP obtained value) which PacRunner handles in an order.
The 'Manual' mode gives entries for different protocols and an entry
for adding no-proxy-for (excludes) for that specific connection.