editor: consider -1 as missing route metric
The route metric can assume a value between 0 and 2^32, or -1 to
indicate that the default value for the device should be used.

In the connection editor we previously saved a missing route metric as
zero and incorrectly displayed the -1 (default) value as 4294967295.

Instead, leave the field empty for default metric. This is similar to
the nmcli convention.

https://bugzilla.gnome.org/show_bug.cgi?id=769015