shared: add "nm-libnm-compat.h"
In the past, when applet used a new API from libnm, we would just
bump the required libnm version. As we anyway released new versions
of the applet with every NetworkManager release, that was ok.

But that means, that 'master' of nm-applet only works with the latest
NetworkManager API. At the same time it means, that a user who is
restricted to an older NetworkManager API cannot use the master build.
Such a user had to rely on important fixes to be backported from master
to the stable branch that was suitable for the NetworkManager API that
he had. That means, upstream there are multiple versions of the applet,
whereas most people would like to have the newest version (or at least
the features and improvements from there).

Now we don't want to branch anymore simply because we require a newer
libnm API. This is somewhat more effort to maintain a master that works
with different libnm APIs, but on the other hand, it saves work by
having to maintain only one upstream branch.

Preferably, we perform runtime detection of the present API. Otherwise,
we do compile-time detection.

"nm-libnm-compat.h" is to privide utilities to access new API or provide
workarounds.