editor: add connection argument to the new page function
The page can use an already existing connection instead of creating a
new one. This would be useful for importing the connections.

Currently we let the new page functions handle the imports (for VPN
connections), but that's not a very good idea -- when we're able to
import non-VPN connections we might not even know the connection type in
advance.

The new page functions are still able to construct the connections
themselves according to details passed and even interact with the user.
That might be worthwhile streamlining at some point.