c-e: rework ethernet link negotiation
Change the autonegotiate CheckButton with a link negotiation ComboBox
allowing the "Ignore", "Automatic" and "Manual" values.
In this way it is possible to remove the "Default" value from the Duplex
and Speed boxes. They will be enabled to enforce the specified values
only when "Manual" is selected.

Move the link negotiation section on bottom.