nma-ui-utils: disable password storage chooser in ask mode
When asking for a password, we cannot modify the password flags. Which means
we cannot modify the password location. The UI gives the wrong impression
that this was possible, while it was not.

Fix that, by disabling the popup menu on the password-storage icon. We
still want to show it so that the user sees it.