all: use gresources
This seems like a good idea. Apart from a tiny speed improvement by
avoiding directory lookups when loading the UI elements and condensing
the XML data by whitespace ommission, this makes it easy to run the
applet and the connection editor without installing them.

This makes it possible to run the binaries directly from the build tree
and not pick up the glade files from whatever happens to be installed in
the system. Neat!