wireless-security: don't reference WirelessSecurity from EAP methods
This causes memory leaks as the WirelessSecurity is never freed due to
circular references: WirelessSecurity references the EAPMethods
through the combo box model, and EAPMethods reference the
WirelessSecurity back.

EAP methods are instantiated by the WirelessSecurity and are supposed
to have a lifetime not longer than the WirelessSecurity's (as they are
hooked in the combo box model and destroyed when the WS is disposed),
so the reference is not needed.