editor: wifi: warn user about unsupported properties
In the Wi-Fi page every time the user saves a connection we replace
the setting with a new one and fill in the properties according to the
UI. This is to ensure that not relevant properties get cleared.

The side effect is also that unsupported properties are not kept after
a save. Warn the user about the properties that will be cleared.