c-e: show error message when --import fails
When --import fails, previously the code would silently proceed and show a dialog
to select a VPN type for creation. Instead an error message should be shown.

The reason the error message was not shown is because at the point when
vpn_connection_from_file() was called, the mainloop was not yet initialized
to show the dialog. Delay the call to vpn_connection_from_file().

It is still broken however, in that nm-c-e would not exit after closing
the error dialog.