build: optionally build with the GCR support
This will be useful for the PKCS#11 token support.

Unfortunatelly, GCR API is declared subject to change despite being in
active use and not having changed in years. This is not a problem --
if GCR ever deprecates this API, chances are it will provide a
certificate chooser too and we'll be just able to drop ours. It doesn't
seem like happening anytime soon.

Let's do an extensive check of how good the available GCR version is and
disable it if it's not good enough. We'll provide a sensible fallback
anyway.