libnma: add a certificate picker interface
A grid with certificate & key pair optionally with passwords.
The idea is to provide reusable interface for the 802.1x and VPN plugins
that can be backed by a simple file-base chooser or a PKCS#11 based
chooser. The API is designed with possible future Gcr certificate chooser
in mind.

This is available from libnm-gtk too, since the wireless-security
library will make use for it. It still resides in the libnma/ tree not
to introduce complexity into how the tree is organized for the sake of a
legacy component.