applet: check permissions before showing 802.1x wifi dialog (CVE-2017-6590)
In most places, we (or NM) check permissions before performing actions.
One place we don't is when we need more information when connecting to
and 802.1x network. In that case we pop up a dialog to ask for more
information before initiaing the connection.

The dialog contains a GTK+ filechooser. We don't want unprivileged users
to have access to this as it allows opening files.

Check for MODIFY_SYSTEM or MODIFY_OWN before showing the dialog for
802.1x connections. If the user doesn't have or can't get it, don't show
the dialog. They wouldn't have been able to create the connection
anyway.

This fixes CVE-2017-6590.

https://mail.gnome.org/archives/networkmanager-list/2017-March/msg00032.html
https://bugs.launchpad.net/bugs/1668321

[bgalvani@redhat.com: changed commit subject line, added links]