cert-chooser: refactor to delegate file/pkcs11 operations from NMACertChooser
Previously, NMACertChooser::constructor() would re-inject the actual
type based on the flags property, thus, instead of creating an instance
of type NMACertChooser, it would create an NMAFileCertChooser or
NMAPkcs11CertChooser.

Don't do that. Don't delegate using inheritence and NMACertChooserClass,
instead delegate using a vtable structure.

Also, hide the actual implementation of NMACertChooser and
NMACertChooserClass from public headers. This means, the classes
cannot be subclassed, but that is something that is not planned
to be supported (at least for the moment).

Co-authored-by: Lubomir Rintel <lkundrak@v3.sk>