applet: fix handling of VPN state changes
vpn_connection_state_changed() is called as handler for the
"state-changed" signal of active connection, but wrongly parses the
argument as NMVpnConnectionState, with the result that we fail to
detect when the VPN becomes active and thus we keep refreshing the
icon animation every 100ms.

Use the correct state values and, while at it, simplify
applet_is_any_vpn_activating() to use the active-connection state
instead of the VPN state.

Fixes: c6e787aac808b2b434455eae8be0137cbab4892a