applet: subscribe to notify::state property of unsupported devices
In different places we call applet_is_any_device_activating() to start
or stop an activation animation. Such fuction takes all devices into
account, including ones not directly supported by the
applet. Therefore, we should update the animation status also when
their state changes.

Without this commit, the applet starts the icon animation when an
unsupported device is activating but does not stop it when the device
changes state.