applet: also react to vpn-state property changes
We use the 'state-changed' signal of active-connection to start and
stop the icon animation and show notifications to user. In
applet_update_icon() we also update the icon and tooltip based on the
VPN state and thus it's necessary to react to changes in the
'vpn-state' property of the vpn-connection, otherwise the icon/tooltip
would be out-of-sync with the real state.