applet: use default active interface, even if it is of unsupported class
Currently if I have bond interface running as my default interface and
I try to establish VPN connection I get route to VPN server added to
physical ethernet interface - one of bond slaves. This makes no sense
- that slave doesn't even have IP address, and obviously such VPN
connection fails to work.

The problem happanes because we try to determine what interface has
default route and we skip all interfaces that we do not know
about (i.e. anything that is not ethernet, wifi, wwan, modem and
BT). Since bond is not on the list code skips it.

It looks like code doesn't really care what type the 'default'
interface is - so that check is really not nessesary. Removing that
check fixes 'VPN with default route via bond' problem.

Signed-off-by: Nikolay Martynov <mar.kolya@gmail.com>

https://mail.gnome.org/archives/networkmanager-list/2017-May/msg00042.html