editor: pppoe: add support for new-style pppoe connections
Add support for the new type of pppoe connections that have an
arbitrary parent device type. When libnm does not support the new
PPPoE 'parent' property, we keep the previous behavior.

When libnm is new enough and we find an old-style connection, its
'interface-name' is used as parent interface and we pick an unused ppp
interface name. In this way, old connections are converted to the new
convention upon save.

https://bugzilla.gnome.org/show_bug.cgi?id=559134