Revert "bridge: add support for group-forward-mask property"
The patch makes use of libnm-1-10 API. At this moment, master of
nm-applet is still supposed to build against 1.8 (current stable).

This should be either resolved by making the use of 1.10 API condtional
(either at compile time or at runtime), or by bumping the required libnm
version. For the moment, revert the patch that breaks build.

https://bugzilla.gnome.org/show_bug.cgi?id=786537

This reverts commit 4a780cc25d69d2dce983a218f5eb435b696a86df.