nma-mobile-wizard: turn into a templated GtkWidget
This makes the mobile wizard intself a GtkAssistant subclass. This
somewhat simplifies the code, but most importantly allows us to use the
GtkBuilder templating to construct it. This simplifies the binding of
widgets that are now described declaratively to the C code.

No more manual construction of the widget hierarchy also means that we
don't use API that is going to be removed in Gtk 4 if we ever decide to
support it.