nma/mobile-providers: read the data files from data directories
Iterate system data directories when looking for country codes and
mobile provider databases instead of hardcoding it to our install
prefix.

This also tolerates when databases are unspecified and missing,
proceeding with an empty provider list. This indicates a broken
installation, but is probably a better thing to do in such case -- the
mobile wizard can still work, just requires the user to specify the APN
name and access technology manually.