build: allow building a GTK4 flavor of libnma
This is EXPERIMENTAL EXPERIMENTAL EXPERIMENTAL. GTK4 is not released yet
and things will change in incompatible ways.

A new library called "libnma-gtk4" is added. The non-libnma parts
(applet, editor) are not ported.

There are few reasons for merging this early and none of them is
shipping the library to the users at this point:

* To allow bringup of GTK4-enabled VPN plugins
* Do not block GTK4 enablement of GNOME Control Center
* Help avoid using features that will be difficult to port from GTK3

All the test programs work, they emit a fair amount of warnings about
properties that were removed in GTK4. To deal with those, we'd need to
build different GResources for GTK3 and GTK4. The differencies seem
small enough to the point the conversion could be automated. TODO,
doesn't block testing.

Tested to build with gtk+-4.0 = 3.94.0, but even the pkg-config name
changed since in GTK4 git master. Did I say this is EXPERIMENTAL and
there will be incompatible changes?