editor: allow creating pppoe connections without parent interface
PPPoE connection that specify a parent interface are more flexible
because they work not only on Ethernet and allow another connection to
be active on the same device. However, due to this last feature,
nm-applet can't control them because it would not be possible to show
two connections active for the same device.

Now the editor can only create pppoe connection with parent-interface
and this made it impossible to activate the connection from the
applet. Add a new checkbox that allows to choose the pppoe connection
type, with a tooltip documenting the behavior of the two choices.

Fixes: 3c83f844ced4f620993054cddb19b22fe459fa0b

https://bugzilla.gnome.org/show_bug.cgi?id=791985
https://bugzilla.redhat.com/show_bug.cgi?id=1574175