build: add GNOME GitLab CI configuration
The pipeline begins with "build" stage doing a distcheck on Fedora 28
(which is still known to ship libnm-glib) and outputting a tarball
artifact.

The output is then used in the "test" stage.

Some repetition is avoided with YAML map capability, but that's fairly
limited.

Both autotools and meson build/installs are done with both autotools and
meson on latest Fedora (happens to be Fedora 28 too). That seems to be a
reasonable enough for start, given we can't test all possible
combinations.

In future, builds on some older platforms, CentOS and Ubuntu and clang
builds would be nice. Not implemented at this point, but it should be
straightforward enough.

Maybe a build with a Git snapshot of NetworkManager would be useful at
some point, but that's not implemented wither.