build: check whether g-ir-scanner actually works
g-ir-scanner uses distutils that have an unfortunate misfeature of
inheriting the compiler flags it itself was built with.

This includes the hardening flags that don't work with without
redhat-rpm-build and break with clang every full moon.

A configure check makes it clear about what went wrong in case
introspection is desired, otherwise turns it off.