c-e: avoid gtk_tree_model_foreach() and iterate tree-model directly
gtk_tree_model_foreach() is not nice in C. The handler callback
is defined far away from where it is called, this makes it harder
to understand what happens.

For example we called,

  gtk_tree_model_foreach (GTK_TREE_MODEL (relabel_list), clear_name_if_present, &filename);

to find out of the relabel_list already contains filename, and if yes,
to return from update_relabel_list_filename(). That is just confusing,
if all we want to do is search relabel_list for a particular filename,
we should do that, and not "clear_name_if_present()".