glade: get rid of GtkContainer.border-width
It's gone from Gtk4. We mostly use it to pad out the pages in the editor
and on a couple other places (where the use is probably incorrect).

The migration manual suggests that a margin of a child should be used
instead. This indeed works for us.

However, we need to set four properties from each side, which is not
nice. An alternative would be to use a CSS class, butthere's no sensible
class in the default theme for the job and it's perhaps not worth adding
a custom style sheet for us.