glade: get rid of GtkContainer.border-width
It's gone from Gtk4. We mostly use it to pad out the pages in the editor
and on a couple other places (where the use is probably incorrect).

The migration manual suggests that a margin of a child should be used
instead. This indeed works for us.

However, we need to set four properties from each side, which is not
nice. An alternative would be to use a CSS class, butthere's no sensible
class in the default theme for the job and it's perhaps not worth adding
a custom style sheet for us.

  https://developer.gnome.org/gtk4/unstable/ch27s02.html#id-1.6.4.4.14

  git grep -l '<property name="border_width">' |xargs sed 's/\(.*\)<property name="border_width">\(.*\)<\/property>/\1<property name="margin_start">\2<\/property>\n\1<property name="margin_end">\2<\/property>\n\1<property name="margin_top">\2<\/property>\n\1<property name="margin_bottom">\2<\/property>/' -i