connection-list: attempt a VPN import first
On nm-connection-editor --import, nm_connection_list_create() is called
with an empty ctype (G_TYPE_INVALID). This is not good -- on a
successful import the first known type (which happens to be Bluetooth)
would match.

Let's treat this as "any connection type" -- attempt a VPN import
opportunistically and if it succeeds, continute with a ctype of
NM_TYPE_SETTING_VPN. Otherwise just bail out.