glade: fix Wi-Fi secrets dialogue being partly untranslated
There was no gettext "domain" property defined for the wifi.ui files,
nor was there a call to gtk_builder_set_translation_domain() before
adding the .ui to the GtkBuilder object, so those strings used the
default project gettext.

Except that when the dialogue is used in another project such as
gnome-control-center, that the default project isn't where the
translations live.

Set the "domain" property on the top-level interface tag, so that
GtkBuilder knows where to find the translations.

[lkundrak@v3.sk: The Wi-Fi dialog is actually composed of widgets from
multiple GtkBuilder files -- set the domain in all of them.]

https://gitlab.gnome.org/GNOME/network-manager-applet/merge_requests/23
https://bugzilla.gnome.org/show_bug.cgi?id=792285