wifi-dialog: remember the specific_connection construct parameter
The specific_connection parameter is passed to
connection_combo_init(), which stores the connection in
priv->connection. When the user changes device, device_combo_changed()
calls again connection_combo_init() passing a NULL argument and so the
specific connection gets lost.

Make this simpler by storing the specific_connection and use it when
the connection combo is re-initialized.