build: replace intltool with gettext
Gettext, since version 0.18, provides infrastructure equivalent to
intltool (generating po/Makefile and translating .desktop files or .xml).

Also, it seems to work better in that it actually honors MSGFMT_OPTS
that are determined by the configure script. That one typically ends up
set to "-c", meaning "check". Without it the bugs in the .po files are
easy to overlook.