applet/vpn-request: rework the auth-dialog communication
This changes the auth-dialog communication in two important ways:

1.) It reworks how is the end of input determined. Previously \n\n whould
    signal a successful end of input, now it checks for EOF instead.

2.) It now no longer reads and processes the input line by line, but
    accumulates the whole thing and only splits it into lines afterwards.

Both changes are done so that external UI mode is easier to add, while
sharing as input handling as possible. The external UI mode uses a GKeyFile
formatted description of the connection secrets. It is parsed as a
whole, not on a line-by-line basis and it is not terminated with a \n\n.

It is functionally requivalent.