applet/vpn-request: fix external UI mode
When a connection is opened for editing in nm-c-e, secrets are asked
to NM. NM in turn asks them to registered agents, including the applet
if available; the applet now supports an external-ui-mode in which it
spawns a VPN-specific authentication dialog binary that doesn't have a
graphical component and contains the knowledge about which secrets are
missing, how to retrieve them from keyrings and how to ask them to
users; the binary then returns through stdout a keyfile that describes
the known secrets, their values returned from keyrings and the labels
to present to users. The applet uses this information to show a GTK
dialog (only when necessary) and returns the secrets to NM.

Currently when using the external-ui-mode and the auth-dialog binary
returns secrets with the ShouldAsk key set to false, the applet
doesn't show a GTK dialog and returns an empty response. This breaks
when the auth-dialog returns secrets from keyrings as they are
dropped.

This commit changes the handling of secrets in external-ui-mode: now
we first build a list of known secrets and, if necessary, populate the
dialog with the ones that must be asked. After the dialog quits we
update the secret values in the list from the dialog and return all
secrets, including ones not asked.

https://gitlab.gnome.org/GNOME/network-manager-applet/issues/59
https://gitlab.freedesktop.org/NetworkManager/NetworkManager/issues/193

Fixes: bce8d0a75c1d ("applet/vpn-request: add external UI mode")